+++ packages/duck-motion/src/anim.ts
@@ -20,10 +20,6 @@ export const AnimVariants = cva('', {
   },
 })
 
-export const AnimDialogVariants = cva(
-  `relative z-50 h-fit w-full border border-border text-popover-foreground bg-popover rounded-lg shadow-sm outline-hidden p-6 flex flex-col overflow-hidden pointer-events-none opacity-0 starting:[&[data-open=true]:opacity-0] data-[open=true]:pointer-events-auto data-[open=true]:opacity-100 text-balance `,
-)
-
 export const AnimPopoverArrowVariants = cva(
   `overflow-visible after:border-background after:w-0 after:h-0 after:absolute after:[position-anchor:var(--position-anchor)] after:[po ition-area:inherit] 
 `,
@@ -49,29 +45,6 @@ after:bg-red-600
     },
     defaultVariants: {
       side: 'left',
-    },
-  },
-)
-
-export const AnimSheetVariants = cva(
-  '!fixed pointer-events-all relative z-50 flex flex-col gap-4 overflow-hidden bg-background p-6 shadow-lg outline-hidden transition ease-in-out data-[open=false]:animate-out data-[open=false]:animate-out data-[open=true]:animate-in data-[open=true]:animate-in sm:max-w-sm',
-
-  {
-    variants: {
-      side: {
-        left: 'data-[open=false]:slide-out-to-left data-[open=true]:slide-in-from-left inset-y-0 left-0 h-full h-screen max-h-screen w-3/4 w-fit border-r border-r data-[open=false]:duration-400 data-[open=true]:duration-300',
-
-        right:
-          'data-[open=false]:slide-out-to-right data-[open=true]:slide-in-from-right inset-y-0 right-0 h-full h-screen max-h-screen w-3/4 w-fit border-l border-l data-[open=false]:duration-400 data-[open=true]:duration-500',
-
-        top: 'data-[open=false]:slide-out-to-top data-[open=true]:slide-in-from-top inset-x-0 top-0 h-3/4 h-fit max-h-screen w-full border-b border-b data-[open=false]:duration-400 data-[open=true]:duration-300',
-
-        bottom:
-          'data-[open=false]:slide-out-to-bottom data-[open=true]:slide-in-from-bottom inset-x-0 bottom-0 h-3/4 h-fit max-h-screen w-full border-t border-t data-[open=false]:duration-400 data-[open=true]:duration-800',
-      },
-    },
-    defaultVariants: {
-      side: 'right',
     },
   },
 )